Extract Java reference helper
